refactored test utils. Added agraph fixtures for unit testing. Backend uses unique fixtures
