+++ tests/testing_utils.h
@@ -6,20 +6,11 @@
 #include <Eigen/Dense>
 
 namespace testutils {
+namespace {
 
 const double TESTING_TOL = 1e-7;
 
-struct AGraphValues {
-Eigen::ArrayXXd x_vals;
-Eigen::VectorXd constants;
-AGraphValues() {}
-AGraphValues(Eigen::ArrayXXd &x, Eigen::VectorXd &c) : 
-  x_vals(x), constants(c) {}
-};
-
-namespace {
-
-inline void print_difference(const Eigen::ArrayXXd &array1,
+inline void print_difference(const Eigen::ArrayXXd& array1,
                              const Eigen::ArrayXXd& array2) {
   std::cout << "  Actual: \n" << array1 << "\n";
   std::cout << "Expected: \n" << array2 << "\n";
@@ -35,8 +26,8 @@ inline double difference(double val_1, double val_2) {
   }
 }
 
-inline bool non_comparable_matrices(const Eigen::ArrayXXd &array1,
-                                    const Eigen::ArrayXXd &array2) {
+inline bool non_comparable_matrices(const Eigen::ArrayXXd& array1,
+                                    const Eigen::ArrayXXd& array2) {
   int rows_array1 = array1.rows();
   int rows_array2 = array2.rows();
   int cols_array1 = array1.cols();
@@ -50,8 +41,8 @@ inline bool non_comparable_matrices(const Eigen::ArrayXXd &array1,
 }
 } // namespace
 
-inline bool almost_equal(const Eigen::ArrayXXd &array1,
-                         const Eigen::ArrayXXd &array2) {
+inline bool almost_equal(const Eigen::ArrayXXd& array1,
+                         const Eigen::ArrayXXd& array2) {
   if (non_comparable_matrices(array1, array2)) {
     print_difference(array1, array2);
     return false;
@@ -73,92 +64,5 @@ inline bool almost_equal(const Eigen::ArrayXXd &array1,
   }
   return equal;
 }
-
-inline AGraphValues init_agraph_vals(double begin, double end, int num_points) {
-  Eigen::VectorXd constants = Eigen::VectorXd(2);
-  constants << 10, 3.14;
-
-  Eigen::ArrayXXd x_vals(num_points, 2);
-  x_vals.col(0) = Eigen::ArrayXd::LinSpaced(num_points, begin, 0);
-  x_vals.col(1) = Eigen::ArrayXd::LinSpaced(num_points, 0, end);
-
-  return AGraphValues(x_vals, constants);
-}
-
-inline std::vector<Eigen::ArrayXXd> init_op_evals_x0(
-    const AGraphValues &sample_agraph_1_values) {
-  Eigen::ArrayXXd x_0 = sample_agraph_1_values.x_vals.col(0);
-  double constant = sample_agraph_1_values.constants[0];
-  Eigen::ArrayXXd c_0 = constant * Eigen::ArrayXd::Ones(x_0.rows());
-
-  std::vector<Eigen::ArrayXXd> op_evals_x0 = std::vector<Eigen::ArrayXXd>();
-  op_evals_x0.push_back(x_0);
-  op_evals_x0.push_back(c_0);
-  op_evals_x0.push_back(x_0+x_0);
-  op_evals_x0.push_back(x_0-x_0);
-  op_evals_x0.push_back(x_0*x_0);
-  op_evals_x0.push_back(x_0/x_0);
-  op_evals_x0.push_back(x_0.sin());
-  op_evals_x0.push_back(x_0.cos());
-  op_evals_x0.push_back(x_0.exp());
-  op_evals_x0.push_back(x_0.abs().log());
-  op_evals_x0.push_back(x_0.abs().pow(x_0));
-  op_evals_x0.push_back(x_0.abs());
-  op_evals_x0.push_back(x_0.abs().sqrt());
-
-  return op_evals_x0;
-}
-
-inline std::vector<Eigen::ArrayXXd> init_op_x_derivs(const AGraphValues &sample_agraph_1_values) {
-  Eigen::ArrayXXd x_0 = sample_agraph_1_values.x_vals.col(0);
-  std::vector<Eigen::ArrayXXd> op_x_derivs = std::vector<Eigen::ArrayXXd>();
-  int size = x_0.rows();
-
-  auto last_nan = [](Eigen::ArrayXXd array) {
-    array(array.rows() - 1, array.cols() -1) = std::nan("1");
-    Eigen::ArrayXXd modified_array = array;
-    return modified_array;
-  };
-  op_x_derivs.push_back(Eigen::ArrayXd::Ones(size));
-  op_x_derivs.push_back(Eigen::ArrayXd::Zero(size));
-  op_x_derivs.push_back(2.0  * Eigen::ArrayXd::Ones(size));
-  op_x_derivs.push_back(Eigen::ArrayXd::Zero(size));
-  op_x_derivs.push_back(2.0 * x_0);
-  op_x_derivs.push_back(last_nan(Eigen::ArrayXd::Zero(size)));
-  op_x_derivs.push_back(x_0.cos());
-  op_x_derivs.push_back(-x_0.sin());
-  op_x_derivs.push_back(x_0.exp());
-  op_x_derivs.push_back(1.0 / x_0);
-  op_x_derivs.push_back(last_nan(x_0.abs().pow(x_0)*(x_0.abs().log()
-                        + Eigen::ArrayXd::Ones(size))));
-  op_x_derivs.push_back(x_0.sign());
-  op_x_derivs.push_back(0.5 * x_0.sign() / x_0.abs().sqrt());
-
-  return op_x_derivs;
-}
-
-inline std::vector<Eigen::ArrayXXd> init_op_c_derivs(const AGraphValues &sample_agraph_1_values) {
-  int size = sample_agraph_1_values.x_vals.rows();
-  Eigen::ArrayXXd c_1 = sample_agraph_1_values.constants[1] * Eigen::ArrayXd::Ones(size);
-  std::vector<Eigen::ArrayXXd> op_c_derivs = std::vector<Eigen::ArrayXXd>();
-
-  op_c_derivs.push_back(Eigen::ArrayXd::Zero(size));
-  op_c_derivs.push_back(Eigen::ArrayXd::Ones(size));
-  op_c_derivs.push_back(2.0  * Eigen::ArrayXd::Ones(size));
-  op_c_derivs.push_back(Eigen::ArrayXd::Zero(size));
-  op_c_derivs.push_back(2.0 * c_1);
-  op_c_derivs.push_back((Eigen::ArrayXd::Zero(size)));
-  op_c_derivs.push_back(c_1.cos());
-  op_c_derivs.push_back(-c_1.sin());
-  op_c_derivs.push_back(c_1.exp());
-  op_c_derivs.push_back(1.0 / c_1);
-  op_c_derivs.push_back(c_1.abs().pow(c_1)*(c_1.abs().log()
-                        + Eigen::ArrayXd::Ones(size)));
-  op_c_derivs.push_back(c_1.sign());
-  op_c_derivs.push_back(0.5 * c_1.sign() / c_1.abs().sqrt());
-
-  return op_c_derivs;
-}
 } // namespace testutils
-
 #endif //BINGO_TEST_TESTING_UTILS_H_
